+++ src/main/java/com/yukgaejang/inflearnclone/domain/social/dto/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.yukgaejang.inflearnclone.domain.dto;
+package com.yukgaejang.inflearnclone.domain.social.dto;
 
 import com.yukgaejang.inflearnclone.domain.board.dto.UserDto;
 import java.util.ArrayList;
