+++ src/main/java/com/yukgaejang/inflearnclone/domain/social/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package com.yukgaejang.inflearnclone.domain.dto;
+package com.yukgaejang.inflearnclone.domain.social.dto;
 
 public interface OAuth2Response {
